// src/app/shared/store/types/navigation.types.ts
export enum NavigationActionTypes {
  NAVIGATE = 'NAVIGATE',
  NAVIGATE_SUCCESS = 'NAVIGATE_SUCCESS',
  NAVIGATE_FAILURE = 'NAVIGATE_FAILURE',
  INITIAL_NAVIGATION = 'INITIAL_NAVIGATION',
  NAVIGATE_BACK = 'NAVIGATE_BACK',
  NAVIGATE_BACK_SUCCESS = 'NAVIGATE_BACK_SUCCESS',
  NAVIGATE_BACK_FAILURE = 'NAVIGATE_BACK_FAILURE',
  NAVIGATE_FORWARD = 'NAVIGATE_FORWARD',
  NAVIGATE_FORWARD_SUCCESS = 'NAVIGATE_FORWARD_SUCCESS',
  NAVIGATE_FORWARD_FAILURE = 'NAVIGATE_FORWARD_FAILURE',
  RESET = 'RESET',
  RESET_CHILDREN = 'RESET_CHILDREN',

  SET_PAGE_VALID_STATUS = 'SET_PAGE_VALID_STATUS'
}